# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Artiv API",
    "description" : "웹툰 플랫폼 백엔드 REST API — 인증(JWT)·작품/회차·뷰어(19금 가드·이미지 정적 서빙)·개인화(구독·읽음·북마크)·소셜(댓글·좋아요)·조회수·성인전용 분류·관리자.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/series/{seriesId}/episodes/{episodeNo}/comments" : {
      "get" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "list_2",
        "parameters" : [ {
          "name" : "seriesId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "episodeNo",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseCommentResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "write",
        "parameters" : [ {
          "name" : "seriesId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "episodeNo",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CommentCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/series/{seriesId}/episodes/{episodeNo}/comments/{commentId}/like" : {
      "post" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "like_1",
        "parameters" : [ {
          "name" : "seriesId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "episodeNo",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        }
      },
      "delete" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "unlike_1",
        "parameters" : [ {
          "name" : "seriesId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "episodeNo",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/posts/{postId}/comments" : {
      "get" : {
        "tags" : [ "post-comment-controller" ],
        "operationId" : "list_4",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PostCommentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "post-comment-controller" ],
        "operationId" : "write_1",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CommentCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostCommentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{postId}/comments/{commentId}/like" : {
      "post" : {
        "tags" : [ "post-comment-controller" ],
        "operationId" : "like_2",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "delete" : {
        "tags" : [ "post-comment-controller" ],
        "operationId" : "unlike_2",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/posts/{postId}/comments/{commentId}/dislike" : {
      "post" : {
        "tags" : [ "post-comment-controller" ],
        "operationId" : "dislike",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "delete" : {
        "tags" : [ "post-comment-controller" ],
        "operationId" : "undislike",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/posts/{id}/like" : {
      "post" : {
        "tags" : [ "post-controller" ],
        "operationId" : "like_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "delete" : {
        "tags" : [ "post-controller" ],
        "operationId" : "unlike_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/posts/{id}/dislike" : {
      "post" : {
        "tags" : [ "post-controller" ],
        "operationId" : "dislike_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "delete" : {
        "tags" : [ "post-controller" ],
        "operationId" : "undislike_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/posts/{id}" : {
      "get" : {
        "tags" : [ "post-controller" ],
        "operationId" : "detail",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostDetailResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "post-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      },
      "patch" : {
        "tags" : [ "post-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PostUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/admin/posts/{id}" : {
      "get" : {
        "tags" : [ "admin-post-controller" ],
        "operationId" : "detail_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostAdminDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/series/{seriesId}/episodes/{episodeNo}/comments/{commentId}" : {
      "delete" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "delete_3",
        "parameters" : [ {
          "name" : "seriesId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "episodeNo",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/posts/{postId}/comments/{commentId}" : {
      "delete" : {
        "tags" : [ "post-comment-controller" ],
        "operationId" : "delete_4",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "IdResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "CommentCreateRequest" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "parentId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "content" ]
      },
...
      "PostCommentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "authorId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "authorNickname" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "likeCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "liked" : {
            "type" : "boolean"
          },
          "dislikeCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "disliked" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "replies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PostCommentResponse"
            }
          }
        }
      },
...
      "PostUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          }
        }
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "sort" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "CommentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "parentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "content" : {
            "type" : "string"
          },
          "authorNickname" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "likeCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "liked" : {
            "type" : "boolean"
          },
          "replies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommentResponse"
            }
          }
        }
      },
      "PageResponseCommentResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommentResponse"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          }
        }
      },
...
      "PostDetailResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "authorId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "category" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "authorNickname" : {
            "type" : "string"
          },
          "likeCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "liked" : {
            "type" : "boolean"
          },
          "dislikeCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "disliked" : {
            "type" : "boolean"
          },
          "images" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PostImageResponse"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PostImageResponse" : {
        "type" : "object",
        "properties" : {
          "url" : {
            "type" : "string"
          },
          "width" : {
            "type" : "integer",
            "format" : "int32"
          },
          "height" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sortOrder" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "PostAdminDetailResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "category" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "authorNickname" : {
            "type" : "string"
          },
          "likeCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "blinded" : {
            "type" : "boolean"
          },
          "blindedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "images" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PostImageResponse"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}